apiserver/params: change retry factor to int64

+++ apiserver/params/internal.go
@@ -648,7 +648,7 @@ type HookRetryStrategy struct {
 	MinRetryTime    time.Duration
 	MaxRetryTime    time.Duration
 	JitterRetryTime bool
-	RetryTimeFactor int
+	RetryTimeFactor int64
 }
 
 // HookRetryStrategyResult holds a HookRetryStrategy or an error.
